requirements.txt
requests
beautifulsoup4
python-dotenv
pydantic
psycopg2
pandas
folium
tabulate